Added Options menu in MainMenu & PauseMenu and some project clean up
