+++ lone-wanderer-game/Content/Content.mgcb
@@ -798,6 +798,30 @@
 /processorParam:TextureFormat=Color
 /build:Sprites/UI/button.png
 
+#begin Sprites/UI/option_bg_9_scale.png
+/importer:TextureImporter
+/processor:TextureProcessor
+/processorParam:ColorKeyColor=255,0,255,255
+/processorParam:ColorKeyEnabled=True
+/processorParam:GenerateMipmaps=False
+/processorParam:PremultiplyAlpha=True
+/processorParam:ResizeToPowerOfTwo=False
+/processorParam:MakeSquare=False
+/processorParam:TextureFormat=Color
+/build:Sprites/UI/option_bg_9_scale.png
+
+#begin Sprites/UI/small_button.png
+/importer:TextureImporter
+/processor:TextureProcessor
+/processorParam:ColorKeyColor=255,0,255,255
+/processorParam:ColorKeyEnabled=True
+/processorParam:GenerateMipmaps=False
+/processorParam:PremultiplyAlpha=True
+/processorParam:ResizeToPowerOfTwo=False
+/processorParam:MakeSquare=False
+/processorParam:TextureFormat=Color
+/build:Sprites/UI/small_button.png
+
 #begin Tilemaps/mountain_landscape_water.png
 /importer:TextureImporter
 /processor:TextureProcessor
