+++ lone-wanderer-game/GameScreens/MenuScreen.cs
@@ -14,10 +14,13 @@ namespace LoneWandererGame.GameScreens
     {
         private new Game1 Game => (Game1)base.Game;
         private Song backgroundMusic;
-        public MenuScreen(Game1 game): base(game)
+        private const float BUTTON_DISTANCE = 80;
+        public MenuScreen(Game1 game) : base(game)
         {
             Vector2 center = Game.WindowDimensions / 2;
             center.Y -= 92;
+            Vector2 buttonPosition = center;
+
             MenuItems = new List<Button>()
             {
                 new Button(Game)
@@ -38,7 +41,17 @@ namespace LoneWandererGame.GameScreens
                     },
                     OnPress = () => {},
                     OnRelease = () => {},
-                    Position = center + new Vector2(0, 80)
+                    Position = buttonPosition += new Vector2(0, BUTTON_DISTANCE)
+                },
+                new Button(Game)
+                {
+                    Text = "Options",
+                    OnClick = () => {
+                        showOptions = true;
+                    },
+                    OnPress = () => {},
+                    OnRelease = () => {},
+                    Position = buttonPosition += new Vector2(0, BUTTON_DISTANCE)
                 },
                 new Button(Game)
                 {
@@ -46,23 +59,45 @@ namespace LoneWandererGame.GameScreens
                     OnClick = Game.Exit,
                     OnPress = () => {},
                     OnRelease = () => {},
-                    Position = center + new Vector2(0, 160)
+                    Position = buttonPosition += new Vector2(0, BUTTON_DISTANCE)
                 }
             };
             foreach (var item in MenuItems)
             {
                 item.LoadContent("Sprites/UI/button.sf");
             }
+            showOptions = false;
+            Vector2 buttonOffset = new Vector2(0, -64 / 2);
+            Vector2 optionsMenuOffset = new Vector2(0, -30f);
+            Vector2 optionsMenuOrigin = center + buttonOffset + optionsMenuOffset;
+            optionsMenu = new OptionsMenu(Game, optionsMenuOrigin);
+
+            optionsBackPressed = false;
+            optionsBackButton = new Button(game)
+            {
+                Text = "Back",
+                OnClick = () => { optionsBackPressed = true; },
+                OnPress = () => { },
+                OnRelease = () => { },
+                Position = optionsMenuOrigin + new Vector2(0f, optionsMenu.getMenuDimensions().Y + BACK_BUTTON_Y_OFFSET)
+            };
+            optionsBackButton.LoadContent("Sprites/UI/button.sf");
         }
 
         public List<Button> MenuItems { get; set; }
 
         private Texture2D background;
 
+        private bool showOptions;
+        private OptionsMenu optionsMenu;
+
+        private Button optionsBackButton;
+        private bool optionsBackPressed;
+        private const float BACK_BUTTON_Y_OFFSET = (64f / 2f) + 10f;
+
         public override void LoadContent()
         {
             base.LoadContent();
-            MediaPlayer.Volume = 0.01f;
             MediaPlayer.IsRepeating = true;
             backgroundMusic = Game.Content.Load<Song>("Sounds/menuchip");
             MediaPlayer.Play(backgroundMusic);
@@ -70,11 +105,49 @@ namespace LoneWandererGame.GameScreens
             ParticleEmitter.Shared.LayerDepth = 0.98f;
         }
 
+        public void windowSizeChanged()
+        {
+            Vector2 center = Game.WindowDimensions / 2;
+            center.Y -= 92;
+            Vector2 buttonPosition = center;
+            foreach (Button item in MenuItems)
+            {
+                item.Position = buttonPosition += new Vector2(0, BUTTON_DISTANCE);
+            }
+
+            Vector2 buttonOffset = new Vector2(0, -64 / 2);
+            Vector2 optionsMenuOffset = new Vector2(0, -30f);
+            Vector2 optionsMenuOrigin = center + buttonOffset + optionsMenuOffset;
+            optionsMenu.centerChanged(optionsMenuOrigin);
+
+            optionsBackButton.Position = optionsMenuOrigin + new Vector2(0f, optionsMenu.getMenuDimensions().Y + BACK_BUTTON_Y_OFFSET);
+        }
+
         public override void Update(GameTime gameTime)
         {
-            foreach(var item in MenuItems)
+            if (Game.SettingsUpdated)
             {
-                item.Update(gameTime);
+                windowSizeChanged();
+                Game.SettingsUpdated = false;
+            }
+
+            if (!showOptions)
+            {
+                foreach(var item in MenuItems)
+                {
+                    item.Update(gameTime);
+                }
+            }
+            else
+            {
+                optionsMenu.Update(gameTime);
+                optionsBackButton.Update(gameTime);
+
+                if (optionsBackPressed)
+                {
+                    showOptions = false;
+                    optionsBackPressed = false;
+                }
             }
             ParticleEmitter.Shared.Update(gameTime);
         }
@@ -85,10 +158,19 @@ namespace LoneWandererGame.GameScreens
             Game.SpriteBatch.Begin(SpriteSortMode.FrontToBack);
             Game.CustomCursor.Draw();
             Game.SpriteBatch.Draw(background, new Rectangle(0, 0, (int)Game.WindowDimensions.X, (int)Game.WindowDimensions.Y), null, Color.White, 0.0f, Vector2.Zero, SpriteEffects.None, 0.8f);
-            foreach(var button in MenuItems)
+            if (!showOptions)
             {
-                button.Draw();
+                foreach (var button in MenuItems)
+                {
+                    button.Draw();
+                }
             }
+            else
+            {
+                optionsMenu.Draw(0.91f);
+                optionsBackButton.Draw(0.93f);
+            }
+
             ParticleEmitter.Shared.Draw(Game.SpriteBatch);
             Game.SpriteBatch.End();
         }
