+++ lone-wanderer-game/LoneWandererGame.csproj
@@ -174,6 +174,9 @@
     <None Update="Content\Sprites\UI\button.sf">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+    <None Update="Content\Sprites\UI\small_button.sf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
     <None Update="Content\Tilemaps\big_map.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
