+++ lone-wanderer-game/Spells/PiercingSpell.cs
@@ -24,7 +24,7 @@ namespace LoneWandererGame.Spells
         {
             base.LoadContent(content);
             sprite.Play("flying");
-            SoundEffect.Play(0.05f, 1.0f, 1.0f);
+            SoundEffect.Play();
         }
 
         public override void Update(GameTime gameTime)
