+++ lone-wanderer-game/Spells/ProjectileSpell.cs
@@ -26,7 +26,7 @@ namespace LoneWandererGame.Spells
         {
             base.LoadContent(content);
             sprite.Play("flying");
-            SoundEffect.Play(0.05f, 1.0f, 1.0f);
+            SoundEffect.Play();
         }
 
         public override void Update(GameTime gameTime)
