+++ lone-wanderer-game/Spells/SpellBook.cs
@@ -3,13 +3,7 @@ using MonoGame.Extended;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using LoneWandererGame.Entity;
-using LoneWandererGame.Enemy;
-using Autofac.Core.Activators.Reflection;
-using System.Security.Cryptography.X509Certificates;
 using MonoGame.Jolpango.Graphics;
 
 namespace LoneWandererGame.Spells
